requirements.txt
requests>=2.25.1
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
numpy>=1.21.0
Pillow>=9.0.0
